gpu: ion: Don't call virt_to_page on addresses from dma_alloc
The virtual address returned from dma_alloc_coherent is not
guaranteed to be valid for use with virt_to_phys. Use the physical
address returned from dma_alloc to calculate the page. This
technically may not work if the handle returned is not a physical
address but this is an acceptable workaround until patches are
integrated into the dma subsystem to generate sg_tables.

Change-Id: I658eab86bf69cc23adabe8ac059a8a960e41bdfd
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>

Conflicts:

	drivers/gpu/ion/ion_cma_secure_heap.c