msm: msm_bus: Fix the type error affecting sum of abs
On legacy chipsets, uint16 was being used to write bandwidth
values to the Fabric registers. However, NoCs support 64-bit integers.
The issue surfaces only when sum of active ab values exceeds the max
ib value in the system, as then the sum of ab values don't impact the
clock anymore. This is exactly what was observed in H-flip video usecase.

This patch fixex the bandwidth sum type due to which clocks were not
updated in those cases where requested ab values are higher than the ib
values.

Change-Id: I701ddb7e7b2d66dd14ef74bc38f3748ac40b76ec
CRs-Fixed: 480365
Signed-off-by: Gagan Mac <gmac@codeaurora.org>