input: bma250: merge ng and motion
Better integration of both drivers with a merge into bma250_ng.
This permit to avoid power control, initialization and release
in both drivers that are not talking together !

eg: avoid power off of bma250 despite it is needed by motion.
    avoid motion to power off bma250 despite it is in use.
    ...

CONFIG_INPUT_BMA250_MOTION needs to bet set to y in your kernel defconfig as previously.