leds: requested_brightness
permit to keep in memory the real brightness requested in case of max_brightness is changed.
This is useful to restore the previous brightness when external software need to limit the max_brightness
during a certain amount of time.

eg: thermanager can request a lower brightness for a while to cool down the device. without this commit when
    the device is cool enough, thermanager restore the max_brightness but the screen brightness is not restored
    and we need to wait a new brightness change to update the brightness of the screen (change by the user, screen timeout with dimmer, screen off/on).

Android GUI doesn't take into account that the brightness can be limited by a dynamic max_brightness settings so this change avoid confusion on Android interface too.