mdp: kcal: fix out of bounds value in lut table
Each color (R, G, B) use 8 bits (0 to 255) but
some well known programs use 257 values (0 to 256).
We only support 256 values on mdp lut.
So to stay compatible we simply accept the value 256
but we store it as 255.

This avoid to have a corrupted screen due to an out of bounds access
with the value 256.